backend/requirements.txt
blinker==1.9.0
certifi==2025.4.26
click==8.2.0
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.0
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.7.1
flask-profiler==1.8.1
gunicorn==23.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
memory-profiler==0.61.0
packaging==25.0
pluggy==1.6.0
psutil==7.0.0
PyJWT==2.10.1
pymongo==4.13.0
pytest==8.3.5
python-dotenv==1.1.0
simplejson==3.20.1
Werkzeug==3.1.3
